Fix variable types.
There is no need for these variables to be size_t because value always
fits 32-bit int.

+++ src/refresh/images.c
@@ -309,7 +309,7 @@ WAL LOADING
 IMG_LOAD(WAL)
 {
     miptex_t    *mt;
-    size_t      w, h, offset, size, endpos;
+    unsigned    w, h, offset, size, endpos;
 
     if (rawlen < sizeof(miptex_t)) {
         return Q_ERR_FILE_TOO_SMALL;
